removed more unused code and added a few unit tests

+++ src/algorithms/imageProcessing/features/HGS.java
@@ -332,7 +332,7 @@ public class HGS {
             //System.out.println(" " + yA + " -- " + yB + " sum="+sum + ", " + sumA + "," + sumB);
         }
         
-        float sim = sum / ((float)Math.min(sumA, sumB));
+        float sim = sum / (Math.min(sumA, sumB));
         
         return sim;
     }
